+++ CHANGELOG.md
@@ -6,10 +6,17 @@ An EtherCAT master written in Rust.
 
 ## [Unreleased] - ReleaseDate
 
+### Added
+
+- [#224](https://github.com/ethercrab-rs/ethercrab/pull/224) Add support for XDP sockets on Linux
+  using the `xdp` feature.
+
 ### Changed
 
 - **(breaking)** [#230](https://github.com/ethercrab-rs/ethercrab/pull/230) Increase MSRV from 1.77
   to 1.79.
+- **(breaking)** [#224](https://github.com/ethercrab-rs/ethercrab/pull/224) `PduRx::receive_frame`
+  now returns `Result<ReceiveAction, Error>` instead of `Result<(), Error>`.
 - [#231](https://github.com/ethercrab-rs/ethercrab/pull/231) Enable reading of up to 64 PDO entries
   per PDO from EEPROM.
 - [#232](https://github.com/ethercrab-rs/ethercrab/pull/232) Use string index from EEPROM to read
@@ -65,8 +72,6 @@ An EtherCAT master written in Rust.
 
 - [#162](https://github.com/ethercrab-rs/ethercrab/pull/162) Add support for FreeBSD and NetBSD
   using BPF.
-- **(breaking)** [#224](https://github.com/ethercrab-rs/ethercrab/pull/224) `PduRx::receive_frame`
-  now returns `Result<ReceiveAction, Error>` instead of `Result<(), Error>`.
 
 ## [0.4.2] - 2024-05-27