+++ CHANGELOG.md
@@ -6,6 +6,12 @@ An EtherCAT master written in Rust.
 
 ## [Unreleased] - ReleaseDate
 
+### Added
+
+- [#234](https://github.com/ethercrab-rs/ethercrab/pull/234) Added
+  `SubDeviceRef::sdo_write_subindices` to more cleanly write multiple SDO sub-indices and prevent
+  mistakes.
+
 ### Changed
 
 - **(breaking)** [#230](https://github.com/ethercrab-rs/ethercrab/pull/230) Increase MSRV from 1.77