+++ CHANGELOG.md
@@ -8,9 +8,8 @@ An EtherCAT master written in Rust.
 
 ### Added
 
-- [#234](https://github.com/ethercrab-rs/ethercrab/pull/234) Added
-  `SubDeviceRef::sdo_write_subindices` to more cleanly write multiple SDO sub-indices and prevent
-  mistakes.
+- [#234](https://github.com/ethercrab-rs/ethercrab/pull/234) Added `SubDeviceRef::sdo_write_array`
+  to more cleanly write multiple SDO sub-indices and prevent mistakes.
 
 ### Changed